dom/base/test/test_bug564047.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=564047
-->
<head>
  <title>Test for Bug 564047</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=564047">Mozilla Bug 564047</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 564047 **/
try {
  document.doctype.appendChild(document.createTextNode("test"));
  ok(false, "Should have thrown an exception");
} catch (e) {
  is(e.name, "HierarchyRequestError");
  ok(e instanceof DOMException, "Should be a DOMException");
  is(e.code, DOMException.HIERARCHY_REQUEST_ERR);
}
</script>
</pre>
</body>
</html>
